backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
ollama 
openai-whisper